+++ delphin/eds/edspenman.py
@@ -39,7 +39,8 @@ def loads(s):
     return xs
 
 
-def dump(es, destination, properties=True, indent=False, encoding='utf-8'):
+def dump(es, destination, properties=True, predicate_modifiers=False,
+         indent=False, encoding='utf-8'):
     """
     Serialize EDS objects to a EDS-PENMAN file.
 
@@ -48,13 +49,16 @@ def dump(es, destination, properties=True, indent=False, encoding='utf-8'):
         es: iterator of :class:`~delphin.eds.EDS` objects to
             serialize
         properties: if `True`, encode variable properties
+        predicate_modifiers (bool): apply EDS predicate modification
+            when *es* are not EDSs and must be converted
         indent: if `True`, adaptively indent; if `False` or `None`,
             don't indent; if a non-negative integer N, indent N spaces
             per level
         encoding (str): if *destination* is a filename, write to the
             file with the given encoding; otherwise it is ignored
     """
-    text = dumps(es, properties=properties, indent=indent)
+    text = dumps(es, properties=properties,
+                 predicate_modifiers=predicate_modifiers, indent=indent)
     if hasattr(destination, 'write'):
         print(text, file=destination)
     else:
@@ -62,7 +66,7 @@ def dump(es, destination, properties=True, indent=False, encoding='utf-8'):
             print(text, file=fh)
 
 
-def dumps(es, properties=True, indent=False):
+def dumps(es, properties=True, predicate_modifiers=False, indent=False):
     """
     Serialize EDS objects to a EDS-PENMAN string.
 
@@ -70,6 +74,8 @@ def dumps(es, properties=True, indent=False):
         es: iterator of :class:`~delphin.eds.EDS` objects to
             serialize
         properties: if `True`, encode variable properties
+        predicate_modifiers (bool): apply EDS predicate modification
+            when *es* are not EDSs and must be converted
         indent: if `True`, adaptively indent; if `False` or `None`,
             don't indent; if a non-negative integer N, indent N spaces
             per level
@@ -78,7 +84,9 @@ def dumps(es, properties=True, indent=False):
     """
     codec = penman.PENMANCodec()
     to_graph = codec.triples_to_graph
-    graphs = [to_graph(to_triples(e, properties=properties)) for e in es]
+    graphs = [to_graph(to_triples(e, properties=properties,
+                                  predicate_modifiers=predicate_modifiers))
+              for e in es]
     return penman.dumps(graphs, indent=indent)
 
 
@@ -89,27 +97,34 @@ def decode(s):
     return from_triples(penman.decode(s).triples())
 
 
-def encode(eds, properties=True, indent=False):
+def encode(eds, properties=True, predicate_modifiers=False, indent=False):
     """
     Serialize a EDS object to a EDS-PENMAN string.
 
     Args:
         e: a EDS object
         properties (bool): if `False`, suppress variable properties
+        predicate_modifiers (bool): apply EDS predicate modification
+            when *eds* is not an EDS and must be converted
         indent (bool, int): if `True` or an integer value, add
             newlines and indentation
     Returns:
         a EDS-PENMAN-serialization of the EDS object
     """
-    triples = to_triples(eds, properties=properties)
+    triples = to_triples(eds, properties=properties,
+                         predicate_modifiers=predicate_modifiers)
     g = penman.PENMANCodec().triples_to_graph(triples)
     return penman.encode(g, indent=indent)
 
 
-def to_triples(e, properties=True):
+def to_triples(e, properties=True, predicate_modifiers=False):
     """
     Encode the Eds as triples suitable for PENMAN serialization.
     """
+    # attempt to convert if necessary
+    if not isinstance(e, EDS):
+        e = EDS.from_xmrs(e, predicate_modifiers)
+
     triples = []
     # sort nodeids just so top var is first
     nodes = sorted(e.nodes, key=lambda n: n.nodeid != e.top)