+++ delphin/main.py
@@ -55,7 +55,6 @@ def call_convert(args):
         args.to,
         properties=(not args.no_properties),
         color=color,
-        pretty_print=args.pretty_print or args.indent is not None,
         indent=args.indent,
         select=args.select,
         show_status=args.show_status,
@@ -175,8 +174,8 @@ convert_parser.add_argument(
     '--from',
     metavar='FMT',
     default='simplemrs',
-    choices=('simplemrs ace mrx mrs-json dmrx dmrs-json dmrs-penman '
-             'eds eds-json eds-penman'.split()),
+    choices=('simplemrs ace mrx mrs-json dmrx simpledmrs dmrs-json '
+             'dmrs-penman eds eds-json eds-penman'.split()),
     help='original representation (default: simplemrs)')
 convert_parser.add_argument(
     '-t',
@@ -192,13 +191,12 @@ convert_parser.add_argument(
     action='store_true',
     help='suppress morphosemantic properties')
 convert_parser.add_argument(
-    '--pretty-print',
-    action='store_true',
-    help='format in a more human-readable way')
-convert_parser.add_argument(
     '--indent',
     metavar='N',
-    help='format with explicit indent N (implies --pretty-print)')
+    nargs='?',
+    default=False,
+    const=True,
+    help='format with explicit indent N')
 convert_parser.add_argument(
     '--color',
     metavar='WHEN',